+++ jvm/selfie-lib/src/jsMain/kotlin/com/diffplug/selfie/guts/WriteTracker.js.kt
@@ -17,7 +17,7 @@ package com.diffplug.selfie.guts
 
 actual data class CallLocation(actual val fileName: String?, actual val line: Int) :
     Comparable<CallLocation> {
-  override fun compareTo(other: CallLocation) =
+  actual override fun compareTo(other: CallLocation) =
       compareValuesBy(this, other, { it.fileName }, { it.line })
   actual fun withLine(line: Int): CallLocation = TODO()
   actual fun ideLink(layout: SnapshotFileLayout): String = TODO()