+++ jvm/selfie-runner-kotest/src/commonMain/kotlin/com/diffplug/selfie/kotest/SnapshotFileLayoutKotest.kt
@@ -25,15 +25,15 @@ import com.diffplug.selfie.guts.atomic
 
 class SnapshotFileLayoutKotest(settings: SelfieSettingsAPI, override val fs: FS) :
     SnapshotFileLayout {
-  internal var smuggledError: AtomicRef<Throwable?> =
+  private var smuggledError: AtomicRef<Throwable?> =
       atomic(if (settings is SelfieSettingsSmuggleError) settings.error else null)
   override val rootFolder = TypedPath.ofFolder(settings.rootFolder.toString())
   private val otherSourceRoots = settings.otherSourceRoots
   override val allowMultipleEquivalentWritesToOneLocation =
       settings.allowMultipleEquivalentWritesToOneLocation
-  val snapshotFolderName = settings.snapshotFolderName
+  private val snapshotFolderName = settings.snapshotFolderName
   internal val unixNewlines = inferDefaultLineEndingIsUnix(rootFolder, fs)
-  val extension: String = ".ss"
+  private val extension: String = ".ss"
   private val cache = SourcePathCache(this::computePathForCall, 64)
   override fun sourcePathForCall(call: CallLocation): TypedPath {
     checkForSmuggledError()
@@ -62,37 +62,42 @@ class SnapshotFileLayoutKotest(settings: SelfieSettingsAPI, override val fs: FS)
     val possibleNames = likelyExtensions.map { "$fileWithoutExtension.$it" }.toSet()
     return fs.fileWalk(folder) { walk -> walk.filter { it.name in possibleNames }.firstOrNull() }
   }
-  fun snapshotPathForClass(className: String): TypedPath {
-    val lastDot = className.lastIndexOf('.')
-    val classFolder: TypedPath
-    val filename: String
-    if (lastDot == -1) {
-      classFolder = rootFolder
-      filename = className + extension
+  fun snapshotPathForClassOrFilename(classNameOrFilename: String): TypedPath {
+    if (!classNameOrFilename.endsWith(".kt")) {
+      val lastDot = classNameOrFilename.lastIndexOf('.')
+      val classFolder: TypedPath
+      val ssFilename: String
+      if (lastDot == -1) {
+        classFolder = rootFolder
+        ssFilename = classNameOrFilename + extension
+      } else {
+        classFolder =
+            rootFolder.resolveFolder(classNameOrFilename.substring(0, lastDot).replace('.', '/'))
+        ssFilename = classNameOrFilename.substring(lastDot + 1) + extension
+      }
+      val ssFolder = snapshotFolderName?.let { classFolder.resolveFolder(it) } ?: classFolder
+      return ssFolder.resolveFile(ssFilename)
     } else {
-      classFolder = rootFolder.resolveFolder(className.substring(0, lastDot).replace('.', '/'))
-      filename = className.substring(lastDot + 1) + extension
-    }
-    val parentFolder = snapshotFolderName?.let { classFolder.resolveFolder(it) } ?: classFolder
-    return parentFolder.resolveFile(filename)
-  }
-  fun subpathToClassname(subpath: String): String {
-    check(subpath.indexOf('\\') == -1)
-    val classnameWithSlashes =
-        if (snapshotFolderName == null) {
-          subpath.substring(0, subpath.length - extension.length)
-        } else {
-          val lastSlash = subpath.lastIndexOf('/')
-          val secondToLastSlash = subpath.lastIndexOf('/', lastSlash - 1)
-          check(secondToLastSlash != -1) { "Expected at least two slashes in $subpath" }
-          check(lastSlash - secondToLastSlash - 1 == snapshotFolderName.length) {
-            "Expected '$subpath' to be in a folder named '$snapshotFolderName'"
+      val lastDot = classNameOrFilename.lastIndexOf(".")
+      val snapshotFileName = classNameOrFilename.substring(0, lastDot)
+      val filename = classNameOrFilename.substring(0, lastDot) + extension
+      val testSource =
+          FSOkio.fileWalk(rootFolder) { seq ->
+            val allPaths = seq.filter { it.name == classNameOrFilename }.toList()
+            if (allPaths.size > 1) {
+              throw AssertionError(
+                  "Cannot store snapshot because snapshot path is ambiguous, multiple files named $classNameOrFilename were found:\n  - ${allPaths.joinToString("\n  - ")}")
+            } else if (allPaths.isEmpty()) {
+              throw AssertionError(
+                  "Cannot store snapshot because no file named $classNameOrFilename was found")
+            } else {
+              TypedPath.ofFile(allPaths.single().absolutePath)
+            }
           }
-          val simpleName = subpath.substring(lastSlash + 1, subpath.length - extension.length)
-          if (secondToLastSlash == -1) simpleName
-          else subpath.substring(0, secondToLastSlash + 1) + simpleName
-        }
-    return classnameWithSlashes.replace('/', '.')
+      val folder = testSource.parentFolder()
+      val ssFolder = snapshotFolderName?.let { folder.resolveFolder(it) } ?: folder
+      return ssFolder.resolveFile(filename)
+    }
   }
 
   companion object {