+++ jvm/selfie-runner-kotest/src/jvmTest/kotlin/com/diffplug/selfie/kotest/StringSpecTest.kt
@@ -15,25 +15,25 @@
  */
 package com.diffplug.selfie.kotest
 
-import io.kotest.matchers.shouldBe
-
-class HarnessVerifyTest : HarnessKotest() {
-  private var initialContent: String = ""
-
+class StringSpecTest : HarnessKotest() {
   init {
-    "initialize" {
-      ut_mirrorKt().restoreFromGit()
-      initialContent = ut_mirrorKt().linesFrom("UT_HarnessVerifyTest").toLast("}").content()
-    }
-    "runAll" { gradleInteractiveFail() }
-    "commentOutFailure" {
-      ut_mirrorKt().linesFrom("alwaysFails()").toFirst("}").commentOut()
-      gradleInteractivePass()
-    }
-    "restoreInitial" {
-      ut_mirrorKt().restoreFromGit()
-      val actualContent = ut_mirrorKt().linesFrom("UT_HarnessVerifyTest").toLast("}").content()
-      actualContent shouldBe initialContent
+    test("smoke") {
+      gradleWriteSS()
+      ut_snapshot()
+          .assertContent(
+              """
+          ╔═ a ═╗
+          a
+          ╔═ b ═╗
+          b
+          ╔═ c ═╗
+          c
+          ╔═ [end of file] ═╗
+
+          """
+                  .trimIndent())
+      gradleReadSS()
+      ut_snapshot().deleteIfExists()
     }
   }
 }